extended handler/scraper
I'm getting a bit in the weeds with the database creation and management.
I'm unsure how the scope of each of these modules should be separated. A lot
of what I'm writing in the master scraper feels like it should be db_handler logic.
Im going to a new branch handler_scope to play around with that idea