checking if exists
Added a ModelMixin class to models with a class method for checking if
entry exists. I see the redundancy of this solution but it is nice and
readable so I kept it. There is also a get or create class method but it
turned out to not serve my purpose as I want to check if exists before
I scrape the data. Other changes to allow for this