check if added
I was having an issue where case_numbers came up multiple times within
a batch. It would pass the 'if not exists_in_table' since it had not yet
been committed. I created a list that holds values of already added keys
and we now check if keys are in this list before adding to session. List
gets reset to empty list on each commit.