handle multiple parcel
New branch to write for handling of fringe cases of data. Occiasionally,
an auction will contain multiple parcels which breaks the adding into auction_lake
table. This commit is a quick patch to bypass the first such case. Eventually
there will be a fuzzy_match function to capture any case where MULTIPLE PARCEL mispellings
are in the auction page. Other known fringe cases are when there is no parcel_id or
folio in the auction page