Beef up models
Instead of a table of just dates I am creating a new data lake with the
html for each calendar date scraped from foreclosure website. Might as well
store that data as well.

Also changed DataLake super class to RequestsDataLake and added new columns
to store response headers and status code. When parsing the data in the future
it might be nice to have these data points available.